Add a `Message` type.
`Message` is a general abstraction for the "unit" of conversational
interaction. It is being added to mainly serves these purposes:

1. Provide a general, event-driven, object-oriented interface for Slack API
messages. In other words, the message is treated as a concrete thing instead of
just an abstract idea. This makes it easy to work with things like reactions,
since it reconciles the RTM and Web API interactions into a single interface.
2. Provide a drop-in replacement for the old `Response` type, whose abstraction
was beginning to leak in many ways.
3. Provide the foundation for a true Conversation type.

This is a major overhaul of the Botstrap internals. Although full API
compatibility is intended, some of the methods previously provided by
`Response` are unimplemented.